pyproject.toml
[project]

name = "idstools"

version = "0.0.1"

description = "A set of helpful tools."

readme = "README.md"

requires-python = ">=3.10"

license = {file = "LICENSE"}

keywords = ["data", "science", "studies", "development"]

authors = [
  {name = "David Rmn", email = "contact@rmn.rocks" }
]

maintainers = [
  {name = "David Rmn", email = "contact@rmn.rocks" }
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Data Scientists",
  "Topic :: Data Analysis :: Data Science",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "pandas",
  "pyyaml",
  "missingno",
  "altair",
  "seaborn",
  "scikit-learn",
  "dynaconf",
  "tqdm"
]

[project.urls]
"Homepage" = "https://github.com/DavidRmn/introduction-data-science"

[project.scripts]
idstools = "idstools:main"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"